add unbonding store

+++ params/params.go
@@ -19,7 +19,7 @@ func NewLocalParamsRetriever() *LocalParamsRetriever {
 	privKey, _ := btcec.NewPrivateKey()
 	covenantPks := []*btcec.PublicKey{privKey.PubKey()}
 	covenantQuorum := uint32(1)
-	unbondingTime := uint32(1000)
+	unbondingTime := uint16(1000)
 
 	return &LocalParamsRetriever{params: &types.Params{
 		MagicBytes:     magicBytes,
